+++ validationframework-core/src/main/java/com/google/code/validationframework/base/validator/generalvalidator/dsl/MultipleResultCollectorContext.java
@@ -101,11 +101,11 @@ public class MultipleResultCollectorContext<DPO> {
 
     /**
      * Makes the validator process each data provider independently.<br>This corresponds to the use of {@link
-     * GeneralValidator.DataProviderToRuleMapping#EACH_TO_EACH}.
+     * GeneralValidator.MappingStrategy#SPLIT}.
      *
      * @return Context allowing further construction of the validator using the DSL.
      *
-     * @see {@link GeneralValidator.DataProviderToRuleMapping#EACH_TO_EACH}.
+     * @see {@link GeneralValidator.MappingStrategy#SPLIT}.
      */
     public ForEachDataProviderContext<DPO> forEach() {
         // Change context
@@ -127,7 +127,7 @@ public class MultipleResultCollectorContext<DPO> {
         }
 
         return new TransformedDataProviderContext<DPO, TDPO>(addedTriggers, addedDataProviders,
-                GeneralValidator.DataProviderToRuleMapping.ALL_TO_EACH, transformers);
+                GeneralValidator.MappingStrategy.JOIN, transformers);
     }
 
     /**
@@ -146,7 +146,7 @@ public class MultipleResultCollectorContext<DPO> {
 
         // Change context
         return new SingleRuleContext<DPO, Collection<DPO>, RO>(addedTriggers, addedDataProviders,
-                GeneralValidator.DataProviderToRuleMapping.ALL_TO_EACH, null, addedRules);
+                GeneralValidator.MappingStrategy.JOIN, null, addedRules);
     }
 
     /**
@@ -165,6 +165,6 @@ public class MultipleResultCollectorContext<DPO> {
 
         // Change context
         return new SingleRuleContext<DPO, Collection<DPO>, RO>(addedTriggers, addedDataProviders,
-                GeneralValidator.DataProviderToRuleMapping.ALL_TO_EACH, null, addedRules);
+                GeneralValidator.MappingStrategy.JOIN, null, addedRules);
     }
 }