+++ validationframework-core/src/main/java/com/google/code/validationframework/base/validator/generalvalidator/dsl/MultipleRuleContext.java
@@ -58,7 +58,7 @@ public class MultipleRuleContext<DPO, RI, RO> {
     /**
      * Data provider to rule mapping to be set to the validator under construction.
      */
-    private final GeneralValidator.DataProviderToRuleMapping dataProviderToRuleMapping;
+    private final GeneralValidator.MappingStrategy dataProviderToRuleMapping;
 
     /**
      * Rule input transformers to be added to the validator under construction.
@@ -81,7 +81,7 @@ public class MultipleRuleContext<DPO, RI, RO> {
      */
     public MultipleRuleContext(Collection<Trigger> addedTriggers, //
                                Collection<DataProvider<DPO>> addedDataProviders, //
-                               GeneralValidator.DataProviderToRuleMapping dataProviderToRuleMapping, //
+                               GeneralValidator.MappingStrategy dataProviderToRuleMapping, //
                                Collection<Transformer> addedRuleInputTransformers, //
                                Collection<Rule<RI, RO>> addedRules) {
         this.addedTriggers = addedTriggers;
@@ -125,11 +125,11 @@ public class MultipleRuleContext<DPO, RI, RO> {
 
     /**
      * Makes the validator process each rule independently.<br>This corresponds to the use of {@link
-     * GeneralValidator.RuleToResultHandlerMapping#EACH_TO_EACH}.
+     * GeneralValidator.MappingStrategy#SPLIT}.
      *
      * @return Context allowing further construction of the validator using the DSL.
      *
-     * @see {@link GeneralValidator.RuleToResultHandlerMapping#EACH_TO_EACH}.
+     * @see {@link GeneralValidator.MappingStrategy#SPLIT}.
      */
     public ForEachRuleContext<DPO, RI, RO> forEach() {
         // Change context
@@ -154,7 +154,7 @@ public class MultipleRuleContext<DPO, RI, RO> {
 
         return new TransformedRuleContext<DPO, RI, RO, TRO>(addedTriggers, addedDataProviders,
                 dataProviderToRuleMapping, addedRuleInputTransformers, addedRules,
-                GeneralValidator.RuleToResultHandlerMapping.ALL_TO_EACH, transformers);
+                GeneralValidator.MappingStrategy.JOIN, transformers);
     }
 
     /**
@@ -173,7 +173,7 @@ public class MultipleRuleContext<DPO, RI, RO> {
         // Change context
         return new ResultHandlerContext<DPO, RI, RO, Collection<RO>>(addedTriggers, addedDataProviders,
                 dataProviderToRuleMapping, addedRuleInputTransformers, addedRules,
-                GeneralValidator.RuleToResultHandlerMapping.ALL_TO_EACH, null, addedResultHandlers);
+                GeneralValidator.MappingStrategy.JOIN, null, addedResultHandlers);
     }
 
     /**
@@ -191,6 +191,6 @@ public class MultipleRuleContext<DPO, RI, RO> {
         }
 
         // Change context
-        return new ResultHandlerContext<DPO, RI, RO, Collection<RO>>(addedTriggers, addedDataProviders, dataProviderToRuleMapping, addedRuleInputTransformers, addedRules, GeneralValidator.RuleToResultHandlerMapping.ALL_TO_EACH, null, addedResultHandlers);
+        return new ResultHandlerContext<DPO, RI, RO, Collection<RO>>(addedTriggers, addedDataProviders, dataProviderToRuleMapping, addedRuleInputTransformers, addedRules, GeneralValidator.MappingStrategy.JOIN, null, addedResultHandlers);
     }
 }